Refactoring brave theme related modules
Enable webui/base native dark mode on platforms that doesn't support
system dark mode like win7/win8/linux/mac high sierra.
