+++ browser/themes/BUILD.gn
@@ -5,8 +5,8 @@ source_set("themes") {
       "brave_theme_service.h",
       "brave_theme_service_win.cc",
       "brave_theme_service_win.h",
+      "brave_theme_utils.cc",
       "brave_theme_utils.h",
-      "brave_theme_utils_linux.cc",
       "brave_theme_utils_mac.mm",
       "brave_theme_utils_win.cc",
       "theme_properties.cc",
