+++ browser/themes/brave_theme_utils.h
@@ -10,11 +10,21 @@
 
 bool SystemThemeSupportDarkMode();
 
-// Override system theme with |type|. With this, browser gets this theme type
-// regardless of OS preference when it queries system theme type.
-// If |type| is BRAVE_THEME_TYPE_DEFAULT, clear overridden theme and follow
-// the theme in OS preference.
-// Note: This is only implemented on MacOS for now.
+// When system supports system per-application system theme changing, set it.
+// Currently, only MacOS support it.
+// Otherewise, we need to overrides from native theme level and explicitly
+// notifying to let observers know.
+// By overriding, base ui components also use same brave theme type.
 void SetSystemTheme(BraveThemeType type);
+
+// Inserted in the ui namespace to add into ui::NativeTheme/NativeThemeWin as a
+// friend function. This function calls protected/private method of
+// ui::NativeTheme::set_dark_mode(). It's a protected method that called by
+// platform specific subclasses whenever system os theme is changed.
+// But we want to change it for using brave theme also for webui/base ui
+// modules like context menu.
+namespace ui {
+void SetDarkMode(bool dark_mode);
+}  // namespace ui
 
 #endif  // BRAVE_BROWSER_THEMES_BRAVE_THEME_UTILS_H_
