+++ patches/ui-native_theme-native_theme.h.patch
@@ -1,0 +1,12 @@
+diff --git a/ui/native_theme/native_theme.h b/ui/native_theme/native_theme.h
+index 15f53916e735a8bd4d7b50d5a06b2117b4d3574a..dc0492f685eb29290db09a2bf61ad20f4228c9b3 100644
+--- a/ui/native_theme/native_theme.h
++++ b/ui/native_theme/native_theme.h
+@@ -428,6 +428,7 @@ class NATIVE_THEME_EXPORT NativeTheme {
+   void SetDarkModeParent(NativeTheme* dark_mode_parent);
+ 
+  protected:
++  friend void SetDarkMode(bool dark_mode);
+   NativeTheme();
+   virtual ~NativeTheme();
+
