--- patches/ui-native_theme-native_theme_win.cc.patch
@@ -1,12 +1,0 @@
-diff --git a/ui/native_theme/native_theme_win.cc b/ui/native_theme/native_theme_win.cc
-index 4cc71daea3f337afaf510e86cdc2f776d377a23c..d9fe7aab59eda7ae0d746cb87ccda3f12686a3b5 100644
---- a/ui/native_theme/native_theme_win.cc
-+++ b/ui/native_theme/native_theme_win.cc
-@@ -590,6 +590,7 @@ bool NativeThemeWin::SystemDarkModeEnabled() const {
-   // ...unless --force-dark-mode was specified in which case caveat emptor.
-   if (UsesHighContrastColors() && !IsForcedDarkMode())
-     return false;
-+  if (OverrideSystemDarkMode()) return GetSystemDarkModeEnabledOverrides();
-   return NativeTheme::SystemDarkModeEnabled();
- }
-
